Renamed files to better follow coding conventions
